# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DistSys slots API",
    "version": "0.4.0",
    "description": "Three roles: admin manages homework numbers and defense dates; assistants publish only on allowed dates; students provide a full name and repository username. Participant names, repository links and Telegram username are returned only to the owning assistant, never by admin calendar endpoints. Verified Telegram sessions; mutations require Origin and X-CSRF-Token. Administrators can use assistant and student endpoints without a separate assistant role; assistants can also use student endpoints. Resource ownership checks still apply. The monthly calendar exposes anonymous counts, not participant identities. Staff display uses only @username from verified Telegram data, or Ассистент if missing; never Telegram first/last names. Roles are assigned through the operator CLI."
  },
  "servers": [
    {
      "url": "https://slots.example.com/api/v1"
    }
  ],
  "security": [
    {
      "session": []
    }
  ],
  "paths": {
    "/auth/telegram": {
      "post": {
        "summary": "Verify Telegram initData and issue a one-hour secure session",
        "parameters": [
          {
            "in": "header",
            "name": "Origin",
            "required": true,
            "schema": {
              "type": "string",
              "example": "https://slots.example.com"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "csrf_token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "400 invalid input; 401 expired session; 403 role/CSRF/origin; 404 inaccessible object; 409 scheduling conflict; 429 rate limit; 500 internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "init_data": {
                    "type": "string"
                  }
                },
                "required": [
                  "init_data"
                ]
              }
            }
          }
        }
      }
    },
    "/me": {
      "get": {
        "summary": "Current profile, roles and CSRF token",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    },
                    "csrf_token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "400 invalid input; 401 expired session; 403 role/CSRF/origin; 404 inaccessible object; 409 scheduling conflict; 429 rate limit; 500 internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update own profile",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Origin",
            "required": true,
            "schema": {
              "type": "string",
              "example": "https://slots.example.com"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "400 invalid input; 401 expired session; 403 role/CSRF/origin; 404 inaccessible object; 409 scheduling conflict; 429 rate limit; 500 internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "full_name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "repository_username": {
                    "type": "string",
                    "minLength": 1,
                    "description": "Suffix of https://distsys.ru/hse-2026/. Whitespace and forward/back slashes are forbidden; other characters are accepted and URL-escaped."
                  }
                },
                "required": [
                  "full_name",
                  "repository_username"
                ],
                "additionalProperties": false,
                "description": "Private full name and repository username. Telegram staff identity is unchanged."
              }
            }
          }
        }
      }
    },
    "/homeworks": {
      "get": {
        "summary": "Active homework catalog",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Homework"
                  }
                }
              }
            }
          },
          "default": {
            "description": "400 invalid input; 401 expired session; 403 role/CSRF/origin; 404 inaccessible object; 409 scheduling conflict; 429 rate limit; 500 internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/slots": {
      "get": {
        "summary": "Future unoccupied slots",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 100
            }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 100000,
              "default": 0
            }
          },
          {
            "in": "query",
            "name": "homework_id",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "schema": {
              "type": "string",
              "format": "date",
              "description": "Inclusive date in Europe/Moscow"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Slot"
                  }
                }
              }
            }
          },
          "default": {
            "description": "400 invalid input; 401 expired session; 403 role/CSRF/origin; 404 inaccessible object; 409 scheduling conflict; 429 rate limit; 500 internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/slots/{id}": {
      "get": {
        "summary": "Slot details without student identities",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Slot"
                }
              }
            }
          },
          "default": {
            "description": "400 invalid input; 401 expired session; 403 role/CSRF/origin; 404 inaccessible object; 409 scheduling conflict; 429 rate limit; 500 internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/slots/{id}/bookings": {
      "post": {
        "summary": "Book for current student; repeat returns existing booking",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Origin",
            "required": true,
            "schema": {
              "type": "string",
              "example": "https://slots.example.com"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "booking_id": {
                      "type": "integer",
                      "format": "int64"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "400 invalid input; 401 expired session; 403 role/CSRF/origin; 404 inaccessible object; 409 scheduling conflict; 429 rate limit; 500 internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/me/bookings": {
      "get": {
        "summary": "Own bookings and cancellation history",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 100
            }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 100000,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Booking"
                  }
                }
              }
            }
          },
          "default": {
            "description": "400 invalid input; 401 expired session; 403 role/CSRF/origin; 404 inaccessible object; 409 scheduling conflict; 429 rate limit; 500 internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/bookings/{id}/cancel": {
      "post": {
        "summary": "Cancel own future booking",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Origin",
            "required": true,
            "schema": {
              "type": "string",
              "example": "https://slots.example.com"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "400 invalid input; 401 expired session; 403 role/CSRF/origin; 404 inaccessible object; 409 scheduling conflict; 429 rate limit; 500 internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/assistant/windows": {
      "get": {
        "summary": "All own windows, newest first",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Window"
                  }
                }
              }
            }
          },
          "default": {
            "description": "400 invalid input; 401 expired session; 403 role/CSRF/origin; 404 inaccessible object; 409 scheduling conflict; 429 rate limit; 500 internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "All own past, future and cancelled windows in one response, ordered by starts_at DESC, id DESC. No pagination or date filters."
      },
      "post": {
        "summary": "Publish one window (assistant)",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Origin",
            "required": true,
            "schema": {
              "type": "string",
              "example": "https://slots.example.com"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id"
                  ],
                  "properties": {
                    "id": {
                      "type": "integer",
                      "format": "int64"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "400 invalid input; 401 expired session; 403 role/CSRF/origin; 404 inaccessible object; 409 scheduling conflict; 429 rate limit; 500 internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WindowInput"
              }
            }
          }
        },
        "description": "Publishes one window and its slots atomically. Overlapping windows are allowed. Future times, allowed defense dates, slot exclusions and idempotency are checked. Repeating the same request with the same Idempotency-Key returns the original window ID."
      }
    },
    "/assistant/windows/{id}": {
      "get": {
        "summary": "Own window with participants (assistant)",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Window"
                }
              }
            }
          },
          "default": {
            "description": "400 invalid input; 401 expired session; 403 role/CSRF/origin; 404 inaccessible object; 409 scheduling conflict; 429 rate limit; 500 internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/assistant/windows/{id}/cancel": {
      "post": {
        "summary": "Cancel own future resource; reason optional even if occupied",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Origin",
            "required": true,
            "schema": {
              "type": "string",
              "example": "https://slots.example.com"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "400 invalid input; 401 expired session; 403 role/CSRF/origin; 404 inaccessible object; 409 scheduling conflict; 429 rate limit; 500 internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "maxLength": 2000
                  }
                },
                "description": "Reason is optional, including cancellation of confirmed bookings. If provided, it is saved and included in student notifications."
              }
            }
          }
        }
      }
    },
    "/assistant/slots/{id}/cancel": {
      "post": {
        "summary": "Cancel own future resource; reason optional even if occupied",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Origin",
            "required": true,
            "schema": {
              "type": "string",
              "example": "https://slots.example.com"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "400 invalid input; 401 expired session; 403 role/CSRF/origin; 404 inaccessible object; 409 scheduling conflict; 429 rate limit; 500 internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "maxLength": 2000
                  }
                },
                "description": "Reason is optional, including cancellation of confirmed bookings. If provided, it is saved and included in student notifications."
              }
            }
          }
        }
      }
    },
    "/public/config": {
      "get": {
        "summary": "Public bot username for launch page",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "bot_username": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "400 invalid input; 401 expired session; 403 role/CSRF/origin; 404 inaccessible object; 409 scheduling conflict; 429 rate limit; 500 internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/admin/homeworks": {
      "post": {
        "summary": "Create unique homework number and defense dates (admin only)",
        "parameters": [
          {
            "in": "header",
            "name": "Origin",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "format": "int64"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "400 invalid input; 401 authentication; 403 wrong role; 404 missing object; 409 conflict (duplicate homework number).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HomeworkInput"
              }
            }
          }
        }
      }
    },
    "/admin/homeworks/{id}": {
      "patch": {
        "summary": "Update homework number and allowed dates without changing existing windows (admin only)",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "header",
            "name": "Origin",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "format": "int64"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "400 invalid input; 401 authentication; 403 wrong role; 404 missing object; 409 conflict (duplicate homework number).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HomeworkInput"
              }
            }
          }
        },
        "description": "Replace the homework number and allowed dates for new publications. Existing windows, slots and bookings remain unchanged when dates are removed."
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "distsys_session"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        }
      },
...
      "Homework": {
        "allOf": [
          {
            "$ref": "#/components/schemas/HomeworkInput"
          },
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "format": "int64"
              },
              "title": {
                "type": "string",
                "example": "ДЗ №1"
              },
              "window_count": {
                "type": "integer"
              }
            }
          }
        ]
      },
...
    }
  }
}